+++ clx_invoice_policy/models/sale_order_line.py
@@ -7,13 +7,30 @@ from datetime import date
 
 from dateutil.relativedelta import relativedelta
 
-from odoo import fields, models, _
+from odoo import fields, models, api, _
 
 
 class SaleOrderLine(models.Model):
     _inherit = 'sale.order.line'
 
     is_invoiced = fields.Boolean(string='invoiced')
+
+    @api.depends('product_uom_qty', 'discount', 'price_unit', 'tax_id')
+    def _compute_amount(self):
+        for line in self:
+            price = line.price_unit * (1 - (line.discount or 0.0) / 100.0)
+            if line.order_id.partner_id.management_company_type_id and line.order_id.partner_id.management_company_type_id.is_flat_discount:
+                price = line.price_unit - line.order_id.partner_id.management_company_type_id.flat_discount
+            taxes = line.tax_id.compute_all(price, line.order_id.currency_id, line.product_uom_qty,
+                                            product=line.product_id, partner=line.order_id.partner_shipping_id)
+            line.update({
+                'price_tax': sum(t.get('amount', 0.0) for t in taxes.get('taxes', [])),
+                'price_total': taxes['total_included'],
+                'price_subtotal': taxes['total_excluded'],
+            })
+            if self.env.context.get('import_file', False) and not self.env.user.user_has_groups(
+                    'account.group_account_manager'):
+                line.tax_id.invalidate_cache(['invoice_repartition_line_ids'], [line.tax_id.id])
 
     def calculate_qty(self, today, num_of_month):
         qty = 0